requirements.txt
llama-cpp-python==0.2.11
SpeechRecognition==3.10.0
pyttsx3==2.90
webbrowser==0.10.6
transformers==4.30.0
torch==2.0.0
pyautogui==0.9.53
keyboard==0.13.5
pytesseract==3.10.0
pillow==9.5.0
opencv-python==4.7.0.72